authentication backend
implemented the sessions backend portion for the authentication. includes using the google api and using a sessions provider to wrap the children component which in this case is the layout. screen now displays a login string rather than the app to wait for user to log in and trigger a new session.